Code Automation: passing the coefficients to Java Script through json, and triggering the server run to server html page
