--- js code/animating_the_drawing.js
@@ -1,85 +1,0 @@
-time = 0;
-drawx = [];
-drawy = [];
-
-let fourierCoefficients = ['-7.146798815306666+7.136669873507912i', '-5.028503166554759+1.2590107951096272i', '-16.991440121252523+45.843249165770104i', '-0.9206824436205516-1.0442713967985797i', '-14.936706433305256+6.352184141147077i', '-20.71601443421162-16.449013702828807i', '-4.544024110612102-12.781081148302034i', '6.330224231071611+1.6186890431369054i', '-15.123986102593374+1.4149467028376825i', '0.3760446334907995+6.362817594207321i', '2.93440154832558-2.6777100345696243i', '5.820682496599738-0.11022421061138196i', '-0.022500622768452508-2.8571126895728205i', '-3.52512473843657+2.571921811682273i', '4.774443378484334-1.0662741285529747i', '0.5599082811086575-3.0315159934607503i', '-0.6107450023637746-0.4516019905835139i', '1.87347843075549+0.35187901159755836i', '0.006429489215343795+1.8117113746870088i', '-0.37063286283461455-2.718369997120627i', '-0.30881696211703125+0.0717287289686496i', '-0.3543355783609619-0.3794067767716751i', '0.7919660036519113+0.8327308423549857i', '-2.408935893254873-0.4310186762648796i', '-0.40367981545604903-0.32743724808833025i', '0.0927054060563959-0.21504322182563035i', '-1.1469424672141828-0.9635224714957593i', '-0.6110505444736483+0.9372712625841391i', '-0.5768811657326369-0.6878960371787639i', '0.19736974804076143+0.12776635916078938i', '0.3832415843050688-0.01277885976894446i', '0.0053605878387779915+1.1571223017286087i', '0.5923383805682374-0.11951002686314871i', '0.002442164940595931-0.663923122249353i', '0.3746335322630426-0.5523376267444806i', '0.6635507879408222+0.13637829087748368i', '0.04233641120277143-0.28682822129008523i', '-0.2198339130906614-0.1680742379312423i', '0.39573311150764046+0.35230757978353416i', '0.3806860788694474-0.037863906202256326i', '0.1963437811039957+0.24087259430710012i', '-0.7214512753245047-0.0598270845405135i', '-0.16577724283364556+0.2069515581075144i', '0.06411533350529887-0.4893389026595687i', '-0.15270535072664398-0.19383070230418314i', '-0.09486386186005921+0.2728513082984986i', '-0.22760294839487044-0.038995372336212815i', '-0.007650544660333644-0.012411889205491317i', '0.2637856281653866+0.12224157336581977i', '0.03249686880628186+0.2155672625665507i', '-0.06325056763571792-0.21528972598321502i', '-0.19358821482136673-0.17055499993121276i', '0.00249337016714668+0.08924638009231628i', '0.08566474759298158+0.03818287848906297i', '0.014255605911047492-0.1614247650851354i', '-0.14110943432290907+0.026884860443532693i', '-0.21868381238584533+0.2272153676990135i', '0.05795170540953275+0.0884697962937141i', '0.19834611026685736-0.12906795682674255i', '-0.1282168315435203+0.13240352003445632i', '-0.1598542445316619-0.1668421377021262i', '0.1710263618523484-0.02883627912654654i', '0.24307987750407384-0.004308592945369685i', '0.01387262837821164+0.13794695078342498i', '-0.012190839765264132-0.115061643523152i', '0.08289120464778305-0.13514006855705676i', '0.10139252081939533+0.1978116080310335i', '0.06634319595704585+0.07790773744457i', '0.07658262199316665-0.1075442403175152i', '-0.04225987578172892-0.237075540565445i', '-0.036111721355265455+0.1993071780819834i', '0.058516575088578576-0.021840135711034075i', '0.07774842516015609+0.023552111532879408i', '-0.25383772530994464-0.0790008357178511i', '-0.22940086819008176+0.07923136469450287i', '0.023660082443849956+0.031649468078570955i', '0.010182327097491818-0.05687275042872407i', '-0.15714712036090844+0.14992869108002355i', '-0.14826248944049014-0.1448177238111854i', '0.10252248639450395-0.01670061360700114i', '0.11046937008205708+0.031223900732586095i', '-0.010507918144411266+0.17139678291886312i', '0.012632161873033447-0.12840520661966506i', '0.03106525586761419-0.12168432117430536i', '0.049367777396503075+0.06147469954555251i', '0.10402391708307163+0.10064123155812368i', '0.07116780975113343-0.11014352076296385i', '-0.04948687960117763-0.10091304136618667i', '-0.008774321009458623+0.11439749068385693i', '0.11423841050075502+0.007547666628810503i', '0.09997036868946453+0.007987924010829692i', '-0.16271613597776602-0.03325707206530224i', '-0.11487746918089345+0.04547047449067011i', '0.07044723115716946-0.07118191373185358i', '0.048958180229810945+0.005341947317095488i', '-0.08090771625752033+0.08139241254211765i', '-0.09425013833444684-0.05694220670501602i', '0.03815578125111252-0.06332950709878121i', '0.06944898311919842+0.07647311602101331i', '-0.009402355567595805+0.09930627805720421i', '-0.021252914261345936-0.09925506033381112i', '-0.039959945935070046-0.09495234915275604i', '-0.013723464091655035+0.08331593107429514i', '0.04389979811174158+0.05970418045947603i', '0.033967883338917655-0.07770095805205472i', '-0.08343842630760222-0.03306998932786317i', '-0.09594621831146695+0.07377198672619964i', '0.061889413670341564+0.041116763381154345i', '0.08680921638847734-0.048360364764457174i', '-0.08094648660411277+0.042773795889542804i', '-0.08672999383857712-0.049220938784639914i', '0.09312851494302729-0.024017851436364764i', '0.10777614236642287+0.018494911973774744i', '-0.022435831715770183+0.06862533352282386i', '-0.03008953640479571-0.06437265492360657i', '0.04369421216476166-0.08122562241332684i', '0.04803021152909722+0.09571250426320624i', '0.021573210498825185+0.06284462021950074i', '0.024859665799429907-0.07053646905989794i', '-0.024362252955364136-0.10985459175685364i', '-0.023063997557695153+0.10355111004732838i', '0.04284239836840818+0.016524632586926997i', '0.04600291969683255-0.016625075012350006i', '-0.116658692670121-0.045412310223019316i', '-0.11348037371169868+0.04637886497882235i', '0.03750654404702111+0.017137025442298708i', '0.036223805421781274-0.02167824505037519i', '-0.08744637642025468+0.0630656442068765i', '-0.08508493537643624-0.062236072454777074i', '0.06400395010716879-0.0229351137511647i', '0.06572716312318098+0.026566615580514923i', '-0.018829740073460083+0.08578362257682129i', '-0.01391138192790034-0.07660426012240318i', '0.019899638553322427-0.07088160597899107i', '0.023992710699249555+0.05742002959404461i', '0.04365374235083554+0.06822628830535438i', '0.03596164623880748-0.07026045562026939i', '-0.030405547444395442-0.06596859896656786i', '-0.020476213944647374+0.0691880385663835i', '0.06610666046032986+0.01741290371984339i', '0.06240988559547058-0.013365603750639561i', '-0.08746807656755573-0.022845380895345314i', '-0.07449977562272515+0.025875866199954287i', '0.05443458603967884-0.026375294707239972i', '0.04814457574944833+0.008171323626319185i', '-0.0584838919407029+0.04383459108923228i', '-0.06192401387437414-0.0370992959318275i', '0.03400100350103545-0.046908073113274204i', '0.043074476145167837+0.0509707941342043i', '-0.013591688359996246+0.0645532224473988i', '-0.017045355376348437-0.0645439121748217i', '-0.014909413540442201-0.0652473265247532i', '-0.00708048349691455+0.06141539679470533i', '0.027156948350509053+0.04887351881963861i', '0.024287553642132018-0.054549679393091924i', '-0.05421174485348135-0.03301978246593074i', '-0.0583312453416263+0.04649874079605079i', '0.047896951248663236+0.02895389735990751i', '0.05656036656962351-0.0315573417773998i', '-0.05839090216272978+0.0185233421051122i', '-0.06037564228499712-0.02076068398170252i', '0.06435264925517835-0.01637398628438807i', '0.06963840957121212+0.014665919202201338i', '-0.02618863503230468+0.04411914999710132i', '-0.029394675969051914-0.04263338726403431i', '0.031755651791665535-0.05640291887659704i', '0.03336652384849446+0.061762290327960795i', '0.0076393376192073065+0.04915663184593552i', '0.008979271169013786-0.05210005497289085i', '-0.01503150102511109-0.0715412616683756i', '-0.014557065468605874+0.06965603358482664i', '0.03076446434109069+0.022706522767556074i', '0.03167251073976443-0.022805404205696572i', '-0.07257279132174277-0.033885714424091426i', '-0.0718091262862107+0.03433174346093854i', '0.03469090314109596+0.013583858605411676i', '0.034492129944587065-0.015073245236830526i', '-0.061300428108604414+0.0342748361978299i', '-0.06038642085560358-0.03400652495032476i', '0.04782896185038278-0.01895666119319049i', '0.04841332640638661+0.020272193113625547i', '-0.02010957052659087+0.055881702349628935i', '-0.018323739120753763-0.05259384412232593i', '0.016410788903578513-0.05090800729062886i', '0.017915738854974557+0.045981307406972066i', '0.02309744010492812+0.051756879795411895i', '0.020102553475888622-0.0523540361862261i', '-0.020983126595943195-0.04978087680073226i', '-0.017177501122068353+0.05091143988589732i', '0.045472770846271215+0.01949084440697888i', '0.04399930647891799-0.017843562236835267i', '-0.05948397938202082-0.01912505397491144i', '-0.05407488503131583+0.02020467903780114i', '0.04379284873483551-0.012015951013912698i', '0.040967706721760956+0.004500267698383756i', '-0.04610674252174978+0.027827827090748977i', '-0.04735304333932002-0.025150693919521197i', '0.029846907454770602-0.03503173360589039i', '0.03367073003576345+0.036887464934618225i', '-0.014603575380093367+0.04744609903073357i', '-0.016027091022110324-0.047451452230925215i', '-0.005513781840135096-0.04979028955244879i', '-0.0022748306332800103+0.0479916147966403i', '0.017939607159831797+0.04067607664029223i', '0.016835447109903937-0.043186609857222225i', '-0.03844612702669437-0.02991677653871313i', '-0.04029328541795725+0.036104382924701334i', '0.03774584857587024+0.023981242809032237i', '0.04185196678600394-0.025293177618470275i', '-0.045587391329771125+0.007938739389036376i', '-0.04650862859385737-0.008976883688485413i', '0.049670788378682815-0.010847417049591527i', '0.05220660947807822+0.010154892065279614i', '-0.02555871359003429+0.03126386827353604i', '-0.027270441907993454-0.03055347161810698i', '0.026396664678501325-0.04233497124334136i', '0.027160025552630987+0.044872990274499897i', '0.0010402247079438903+0.039845176991168046i', '0.001712594760011378-0.041259808345069426i', '-0.00922546792954429-0.053564042690469965i', '-0.008984040562008123+0.0528547759452213i', '0.022741307913491174+0.02335946132033884i', '0.023061442688158976-0.023412059472745064i', '-0.05129977192118562-0.028319153865197733i', '-0.05107969436764137+0.02853690081422948i', '0.030333430181713014+0.012642156481720887i', '0.030314547960786288-0.013283713297517666i', '-0.04766070010586988+0.020499678758826488i', '-0.047229644417236843-0.02036964151910811i', '0.0390242332813569-0.014680848143055415i', '0.0392522892257283+0.015250645578795088i', '-0.01996937078555109+0.04053978259963022i', '-0.019125270003884625-0.03902326233496782i', '0.015124191983303503-0.03972028741791739i', '0.015834702540257914+0.03741526644645063i', '0.012974552324111821+0.041882113205098784i', '0.011466123538307942-0.042040736650473864i', '-0.014937341977480666-0.040611914326956766i', '-0.013105995680739602+0.04107665627819959i', '0.033804292355494064+0.019717677140055914i', '0.03307295895241621-0.018868904502808953i', '-0.04479525514069235-0.017486862059869605i', '-0.04195678990246088+0.017920321113903695i', '0.03656403706284585-0.005071635722360444i', '0.0349572822168909+0.0012375787551706374i', '-0.038464775893886145+0.018830470677468805i', '-0.03899031591757809-0.017551559084618i', '0.02683633530912756-0.02694923323144126i', '0.028822060731239024+0.027999951278169094i', '-0.014960236841216515+0.03712496665287629i', '-0.01566747978077564-0.03713710231471701i', '-0.0005891905655524693-0.040284329161619434i', '0.000998764531381549+0.03925807625148414i', '0.012035897675891388+0.034916111992240303i', '0.01155269110494193-0.036267967865506484i', '-0.028696754577888173-0.02718880416242084i', '-0.029671633198840117+0.03061932932775352i', '0.030556742236068177+0.02142292188986746i', '0.03288248369521663-0.022235039141152754i', '-0.03731431871497301+0.002148638854125577i', '-0.03782688650957011-0.002718075294451311i', '0.04079402417921625-0.006841711945754222i', '0.042238863889077854+0.006516008905312065i', '-0.02421070629136403+0.023229688955615932i', '-0.025276275289905236-0.022813181396361043i', '0.02345347530694873-0.033280142008442724i', '0.023871983752411098+0.03467517616201951i', '-0.0027632374217519406+0.03326643108735815i', '-0.002375629445212482-0.034050539835872265i', '-0.005235447154018011-0.04316182333697531i', '-0.005079401191698394+0.04287504506919554i', '0.017102723026771974+0.02269631350868083i', '0.01721733437485033-0.02271817575012844i', '-0.03881523179586824-0.025101272670785192i', '-0.03876440756126393+0.025216924938064447i', '0.026383778536851685+0.012484576326693732i', '0.026402772494316065-0.01280426193730401i', '-0.03923634990445806+0.012533603268074622i', '-0.039007745884472644-0.012450290498735028i', '0.033468402888261195-0.011083363241944567i', '0.033550001783579336+0.01135268443148428i', '-0.019499380154561402+0.031099777684790983i', '-0.01903312186395433-0.030288604973211217i', '0.014618560172745224-0.03239331934101459i', '0.015008847118081464+0.031144725254706664i', '0.006970663231046392+0.03526578866627005i', '0.006085911897378355-0.03525481601298315i', '-0.010632419650911579-0.03471061618918089i', '-0.009625772472790003+0.03490701854708111i', '0.026199638427925734+0.019412102437629725i', '0.02578652506021453-0.018905517288199503i', '-0.035637666991562594-0.016666095688743457i', '-0.03392036163117461+0.016826419936198635i', '0.03134634877652771-0.0009359255409046516i', '0.03028552128826921-0.0012885919160480624i', '-0.03327770068095849+0.012988839979758256i', '-0.03351108944503174-0.012314582501831789i', '0.024627156243058237-0.021163258647895734i', '0.025810119266619802+0.021843179352910677i', '-0.01511585609969425+0.0301381447011998i', '-0.015510260367602752-0.030153959615874318i', '0.002485005939107804-0.03379655349611446i', '0.003340483264066125+0.033136050304232355i', '0.007886047977463229+0.030704296134825837i', '0.0076668300756324095-0.031535708039523505i', '-0.022058217110793504-0.025024220360013623i', '-0.02262088425977818+0.02717666405933765i', '0.02524035549507618+0.01991098750696605i', '0.02671710718834417-0.020487844045011833i', '-0.03148272709458025-0.0014723051865550216i', '-0.03180518474766233+0.001124795169532511i', '0.034826028693991457-0.0038161497132179895i', '0.03575072147320661+0.003647819023315471i', '-0.022845314920891632+0.01766502453169093i', '-0.023580153565758374-0.017381249339552027i', '0.021622223482856903-0.026928491496089932i', '0.021874619872374143+0.027775165313530296i', '-0.0052304505740067495+0.028371253739041716i', '-0.004984178733989354-0.028850340001535607i', '-0.0022984391201845214-0.03635179674726247i', '-0.0021787395885074032+0.036244581830012695i', '0.012909186359585573+0.021712645063659206i', '0.01293942566391385-0.02171533178832863i', '-0.03057773274920774-0.02301892159499072i', '-0.030589731247277194+0.023085199579607446i', '0.023037859153198587+0.012598390863330505i', '0.023064775431758715-0.01277091148111098i', '-0.033465886004243516+0.007356135578013519i', '-0.03333504147035404-0.007291931089448686i', '0.029614727817719788-0.00813135441304736i', '0.029625289457089334+0.008262277647580557i', '-0.018972262427551623+0.024632587551643714i', '-0.018687201047652112-0.02415524482593262i', '0.014436897587471631-0.027137374986165648i', '0.01467433959858022+0.026392454434662413i', '0.002982940335761859+0.030481850416025774i', '0.002411205701430658-0.030395608960228068i', '-0.007366266403196806-0.030571683713567978i', '-0.00676545460823531+0.030643680625940042i', '0.020785687533123043+0.018950527620805707i', '0.02053256576767825-0.018616301033671123i', '-0.029304797611632176-0.016221749366434397i', '-0.02816181945364101+0.016247057967852085i', '0.02737920114513493+0.001849684048684297i', '0.02660474047329656-0.0032532636588323144i', '-0.02950330851024619+0.008839697894377822i', '-0.029602711381922296-0.008467991709030842i', '0.02294214334092496-0.016805072862194036i', '0.023720812480902968+0.017285662824900253i', '-0.015193755787563663+0.025036894702959354i', '-0.015431398200527392-0.025054922656312054i', '0.004618894164032737-0.02904145910711771i', '0.005101783168616112+0.02858152854482359i', '0.0047757368094369645+0.027483625416815836i', '0.004682354826659439-0.028047664936440564i', '-0.017215200394066404-0.02330119328257494i', '-0.017553681865608923+0.0247746297972872i', '0.021133457913155677+0.01893010030471999i', '0.022146872426554946-0.01938073638109812i', '-0.027109056145953887-0.003948189371883314i', '-0.027330856543469523+0.003720389870527571i', '0.03050915325306442-0.0014378625296995393i', '0.031152818070522316+0.0013443740381850339i', '-0.02160559551977631+0.013535291425224076i', '-0.022149575091816337-0.013318518942274121i', '0.020381352035695494-0.022188380442227817i', '0.02054390289595969+0.022739675152743895i', '-0.006964316415204702+0.024562384311081327i', '-0.006796279879287364-0.024876641571413824i', '-2.3591830327455987e-05-0.03151220708610937i', '7.919598326460077e-05+0.03148916215861304i', '0.009642825619111436+0.020691448105202376i', '0.009635259239938848-0.020681778466860923i', '-0.024696889359679973-0.021562075091834262i', '-0.02473436030403628+0.02160295907285125i', '0.020212171962412734+0.01281062823924712i', '0.020240006200351983-0.01290711525617505i', '-0.029224074737513928+0.003711791916357593i', '-0.0291450204918351-0.0036564778816850912i', '0.02675888353273214-0.005681858903075848i', '0.026730971856961248+0.005743674816603386i', '-0.01846348178667149+0.019871500009839733i', '-0.018276603342561166-0.01957183148236319i', '0.014404434179743839-0.023128559817725873i', '0.014559964621735626+0.02265305187735555i', '0.0001229093616426022+0.026825415820187222i', '-0.0002706653390413083-0.026702976960110023i', '-0.004772697850356736-0.02748386717446083i', '-0.004395780104537548+0.027492944599938188i', '0.01668752087762823+0.018456214506967376i', '0.016524290779690917-0.018219241839731698i', '-0.024607871520192705-0.015971243623146857i', '-0.02379471535644513+0.01592201738716651i', '0.024233620710909017+0.0038838982938662267i', '0.02362720444161885-0.004820435061669973i', '-0.026609185555748857+0.005703509780382997i', '-0.026642621693252797-0.005499453568439547i', '0.02160810618657452-0.013379652638150073i', '0.02216178139818743+0.013740454115184639i', '-0.015236660720851364+0.021104488969198053i', '-0.01538787111443729-0.02112394744528477i', '0.0062095946011422876-0.0253691414415022i', '0.006483760207762716+0.025031511759212248i', '0.002330906909292163+0.02492339838250776i', '0.0023017340931745546-0.025336347964938014i', '-0.013494672333156883-0.02189985933183497i', '-0.013698023198126113+0.022975080276816037i', '0.017840185439162798+0.018249432694892903i', '0.01857452086390792-0.01862564697249605i', '-0.02367187796803047-0.005753421165617778i', '-0.023835093225972756+0.005596671831054639i', '0.027213757727154293+0.0004967927575898201i', '0.027691047762781373-0.000552571283145814i', '-0.020505617702573274+0.010312159617399834i', '-0.020929056306552767-0.010131757541633257i', '0.019486453292870264-0.01848087967652498i', '0.019596108380202218+0.01885904000884138i', '-0.008256081542486873+0.021487398589515337i', '-0.00813508119873396-0.02170503738384738i', '0.0018095941552135411-0.02786390453962307i', '0.0019041435513041643+0.027882319885667382i', '0.007001671490124144+0.01971646111941354i', '0.006976626126557967-0.0196984779304831i', '-0.02025153212131961-0.020482494044815384i', '-0.020299448346246626+0.020509796815463353i', '0.017795835811061027+0.013055968263559196i', '0.01782316737434315-0.013109684333164603i', '-0.02594022132711967+0.0009926076035860487i', '-0.025890333430366095-0.0009418453216740175i', '0.02453558690279241-0.003610509502256297i', '0.024485156950484192+0.003636336375500468i', '-0.01799192084722563+0.01617847855178376i', '-0.01786323296534562-0.01598178560467102i', '0.014445693994802292-0.019929889563060008i', '0.014553438796205603+0.01961112707844752i', '-0.0020464768803529487+0.023909243650752365i', '-0.0023286032281858472-0.023768734737407538i', '-0.00263903238916674-0.025069799164723572i', '-0.002396632840733052+0.025045401526678973i', '0.013440189525879637+0.017971369430483455i', '0.013331820607913509-0.017794152631138163i', '-0.020941969778143707-0.01582917738501341i', '-0.020335360542195324+0.015735224631767006i', '0.021652717580384888+0.005457123949951528i', '0.021153652670593222-0.006105096205184915i', '-0.024297348007516845+0.0032207271696686607i', '-0.024297340154414416-0.003116874644392531i', '0.02051717097393515-0.010591283486233193i', '0.020936007518161628+0.010873698085726316i', '-0.015262002970676307+0.01794485619182537i', '-0.015362105031639279-0.017965289191399864i', '0.007458224917504176-0.022416066989595022i', '0.007606736681072018+0.022159304394062697i', '0.0003363620367405687+0.022820644133388496i', '0.0003410137622728189-0.023141893792050797i', '-0.010518689041007882-0.020732501007641714i', '-0.010633840568539842+0.02155575133977955i', '0.015115143779655685+0.017752379348744546i', '0.015668076162396353-0.01808144355614605i', '-0.020869388347065997-0.007135429046703966i', '-0.020995910749165424+0.007024093141182311i', '0.024591053726722455+0.002116988947382492i', '0.024962646767457703-0.0021534066569983273i', '-0.01952883751272671+0.007696842738142862i', '-0.01987055480489618-0.007537097060425199i', '0.018809480244123856-0.015471408524954398i', '0.018885897636282865+0.015741308808155323i', '-0.009262601458150381+0.018927537466617947i', '-0.009171793658619712-0.019084886197068426i', '0.003334930395619745-0.024987423173046808i', '0.0034254458788319647+0.025026326742493603i', '0.004799172484568487+0.018808019601898637i', '0.0047663319961423894-0.01878421268128104i', '-0.01674095674325235-0.019646105195218394i', '-0.01679269660801954+0.019666030954440528i', '0.015695306851829925+0.013308324966257667i', '0.015722149306323194-0.01333646814958607i', '-0.023294138640075797-0.001129254283150291i', '-0.02326130153498364+0.0011774960504408738i', '0.022736787015287342-0.0018237951249133605i', '0.022672415344303597+0.0018308012218181818i', '-0.017558836585553406+0.013196584030352718i', '-0.017467111621240882-0.013063580738271338i', '0.014525180639567627-0.01728626403795039i', '0.01460316176452012+0.01706478291736533i', '-0.003764345463782917+0.021503035188366992i', '-0.003971369074083884-0.0213535743045494i', '-0.0008324033141665285-0.023111287098427462i', '-0.0006764325236617366+0.02306873861214931i', '0.010773043817896455+0.01750925213010486i', '0.010700393765867615-0.01737130440180437i', '-0.017966066823954893-0.015751847787568912i', '-0.01749803653630767+0.015629208659727403i', '0.01947397250808032+0.006727560202659313i', '0.019048261374012-0.00718516937010652i', '-0.022388555555739355+0.0011827241246835374i', '-0.022371893113963887-0.0011423898044567171i', '0.019599912677912838-0.008254027030983525i', '0.019933163957397807+0.008481545463157408i', '-0.015277786858290184+0.015320766791995996i', '-0.01534602088955309-0.015341890170090217i', '0.008477889312156185-0.019962897835954563i', '0.00854639125187161+0.019763165391420104i', '-0.001340749454100825+0.021045212568040483i', '-0.0013183999919997302-0.021307606633438017i', '-0.008059185415869887-0.01973764935911472i', '-0.008112948555679518+0.02039141639135987i', '0.012799388354089329+0.017374566778689676i', '0.013226939282224397-0.01767200307727325i', '-0.018514688694105337-0.008235019325637513i', '-0.01861681741786455+0.008154508706717016i', '0.022432570100968947+0.0035080718644516645i', '0.02273308638223112-0.003534840950266585i', '-0.018653704228798855+0.0055071690553772645i', '-0.01893673089200827-0.005359414222337516i', '0.018277530214776093-0.012953283722489245i', '0.018331945784315642+0.013151912225106595i', '-0.010075448971612947+0.016740398052665542i', '-0.01000509907927314-0.016858197139083625i', '0.004638442292056549-0.02263722902521246i', '0.004727082658796816+0.022685607796232805i', '0.002914094671286567+0.017965185498588332i', '0.002878359575135943-0.017937152840952306i', '-0.013870098067267609-0.01897460921654759i', '-0.013922569406040489+0.01899058158987158i', '0.013838152822605618+0.013557356334157278i', '0.013864844750563425-0.013569523138789452i', '-0.021092126989973724-0.002845349398024097i', '-0.02106941636896975+0.0028920908756238095i', '0.021235259404215515-0.0002531239731146384i', '0.02116189578194766+0.0002507696252421754i', '-0.017160087264669898+0.010710070779851796i', '-0.017093191526628547-0.01061854632540142i', '0.01462519940769198-0.015037891878052406i', '0.014683636562333652+0.014879999811428304i', '-0.005172308714562219+0.01946107807623612i', '-0.005325549125579421-0.01930752859849874i', '0.0007349029947393629-0.02147350334450135i', '0.0008321925204344896+0.021421296696856086i', '0.008517354201387443+0.017072152779008185i', '0.008469217315140581-0.016961477962851135i', '-0.015472725305993775-0.015715320316305376i', '-0.01510312862337173+0.015573228381092472i', '0.017589583704473003+0.007789075615463825i', '0.017216845307108167-0.008114109145771153i', '-0.02076857279477223-0.0005402548013725i', '-0.020744429320456587+0.0005383651778757131i', '0.01880994135429584-0.006245432303614892i', '0.019086297356204332+0.006432298664233345i', '-0.015288032651617163+0.013081040313100104i', '-0.015335556074415266-0.013102667290501135i', '0.009337354815641158-0.017869212262926475i', '0.00935254132824944+0.01771188570524141i', '-0.0027871349140244604+0.019510085556277798i', '-0.0027561389165732322-0.0197329446725822i', '-0.005970414256458386-0.018871983719403095i', '-0.005979095048597956+0.019405768555854894i', '0.010785652056852817+0.017077943941940052i', '0.01112198445305495-0.017353141702573502i', '-0.016485609333511075-0.00913788920071315i', '-0.016570678712333288+0.009079388836297012i', '0.02060604452142231+0.004728639831418163i', '0.020856434112237705-0.004751171123313441i', '-0.017860367429921416+0.0036253709752027824i', '-0.018099160362378533-0.0034845512385215383i', '0.01784632741687272-0.010791809675736175i', '0.017885535415178677+0.01094147744377899i', '-0.010751540531799726+0.014829293119067596i', '-0.01069571202243174-0.01492004854473325i', '0.005778117341219018-0.020660011052363356i', '0.005866032776837324+0.02071178905438989i', '0.0012641361199549946+0.017180150078235853i', '0.001228085644514959-0.017148957109144908i', '-0.011453826000928516-0.018419395822315443i', '-0.011505541527470131+0.018433374125409303i', '0.012169158188378762+0.013799461675984004i', '0.01219605854447435-0.013801337408667423i', '-0.019209886161080724-0.004274827236120118i', '-0.01919313396231645+0.004320606475043953i', '0.019948701847414218+0.0011521620264228066i', '0.0198693686719165-0.0011584130960458072i', '-0.01679022990861114+0.008580459946024202i', '-0.016740812951188823-0.008517125899373972i', '0.014736621301153415-0.0130790202203791i', '0.014781641527957821+0.012964537922773557i', '-0.006359561473048135+0.017686428511426128i', '-0.006472265303404258-0.01753149590115788i', '0.0021234472477501074-0.02006856938067087i', '0.0021789670475600292+0.020011645910622544i', '0.0065620628267866726+0.01665816042706346i', '0.00653151361720513-0.016567340973855954i', '-0.013328057417393614-0.015705794761353815i', '-0.013031868603009762+0.015549852478885316i', '0.015925049489715+0.008701169158418953i', '0.015592397956637902-0.008929477673305049i', '-0.019361034068871446-0.0020334647502872755i', '-0.01933463853095623+0.002002360650705165i', '0.018115122618448515-0.004481505058261212i', '0.018352162172840164+0.004636815617112357i', '-0.015294907556974928+0.011124390611897017i', '-0.015328517971186333-0.011146392831865734i', '0.010081058080846507-0.016040655302900268i', '0.010059328385940772+0.015916299488780913i', '-0.004062039272020083+0.01815483520197657i', '-0.004027700914588872-0.018350092943984188i', '-0.004154728853516683-0.01810437009569095i', '-0.004128742759626901+0.018549367621099905i', '0.008998801070057865+0.016838767201932475i', '0.009265779837372382-0.017097688129870933i', '-0.01469862256445238-0.009899069914603778i', '-0.01477121594583733+0.009856994835727245i', '0.019023385035503864+0.005820370980830672i', '0.019236931359476586-0.005841796134916102i', '-0.01713203434236539+0.001971594108842928i', '-0.01733608108041674-0.0018346233329562533i', '0.017487553840715667-0.008895205801131845i', '0.01751585123528302+0.009009987953228701i', '-0.011328083377850979+0.013126069885841044i', '-0.011282998927832556-0.013197656867375444i', '0.006794657354618263-0.01895494083641537i', '0.0068824708805588775+0.01900664330481696i', '-0.00020875763330060722+0.016443395181400908i', '-0.00024373882999772362-0.016409776602216278i', '-0.009369958531645359-0.017948757305070083i', '-0.009420204450187417+0.017961889855267414i', '0.01064592593544937+0.014034045939304407i', '0.010673339750255778-0.01402916561760393i', '-0.017563716142723763-0.005495638567935682i', '-0.017550303946872285+0.00554074739379893i', '0.018821346544987096+0.002430149743185367i', '0.018737976449435104-0.002437023900862291i', '-0.016443892462988434+0.006714421382392394i', '-0.016407291095910985-0.0066709627590083795i', '0.014854703360293025-0.011336416786829798i', '0.014890144782900593+0.011252717687367714i', '-0.007385157873184929+0.016111797770560363i', '-0.00746593389267354-0.015957030076896604i', '0.003376686462751807-0.018836608730354525i', '0.0034011980533918892+0.018777955974342064i', '0.0048307101331040115+0.01626384767611168i', '0.004813279445368475-0.01618809554537333i', '-0.011441362592717792-0.015715008023647267i', '-0.01120230847734291+0.01554875624552389i', '0.014426968987717403+0.009503620362726006i', '0.014125909748107436-0.009658390929256522i', '-0.018112768361158577-0.0033555241187628315i', '-0.018087195254719662+0.0033034819826804687i', '0.01749243812016961-0.002902497856176925i', '0.01770137285027871+0.0030322985626872745i', '-0.015299649127652276+0.009380035689704858i', '-0.015323653073588475-0.0094023214765565i', '0.010739180074886109-0.014411166382371728i', '0.010691008039227186+0.014313450933237597i', '-0.005207572359758647+0.016936024271103578i', '-0.005173106845064147-0.017111341778656343i', '-0.0025440201437458452-0.017411889029251518i', '-0.0024902774516780343+0.017788528168111876i', '0.007384386545221125+0.01664155686343627i', '0.007596545932232468-0.016888136016584544i', '-0.013094357173799719-0.010555497543364947i', '-0.013157444508351615+0.010526175765236366i', '0.01762345825427985+0.006813899271761889i', '0.017808900109381423-0.006836081377133048i', '-0.016454808364241038+0.0004894428731019446i', '-0.016630553800603007-0.000354365689257341i', '0.01718231096730871-0.007198478619757143i', '0.017202513864327675+0.0072875763055335885i', '-0.011830444146599155+0.011581020233354125i', '-0.01179362010878584-0.011638570113961917i', '0.00771767995987222-0.017452767744420547i', '0.007805725577601341+0.017502357568357954i', '-0.0015469845091197094+0.01574551538236424i', '-0.0015801552207934735-0.015709992997323716i', '-0.007534367642669638-0.017541142889927752i', '-0.007582818590810211+0.017554108939918656i', '0.00923534326589629+0.014261882113909899i', '0.00926351424238242-0.014252533306784283i', '-0.01609489980525084-0.006560975065193597i', '-0.01608312169771553+0.006605572007105563i', '0.017813873868895054+0.0036101323423005233i', '0.017727751476528178-0.0036156101525156026i', '-0.01611617693602517+0.005046393679184288i', '-0.016089322678787682-0.005017420538651653i', '0.014977070958482988-0.009757339684422953i', '0.015005421158338215+0.009696249724404256i', '-0.008289870415377869+0.014688708552886398i', '-0.00834450782118494-0.014535004523347501i', '0.004526914614921027-0.017735230931688224i', '0.0045274976234014955+0.017676711426091674i', '0.0032685646816321516+0.01588535016447852i', '0.0032612683921223254-0.01582146978196288i', '-0.0097486069495156-0.01573789093297734i', '-0.009555760679391817+0.015563636227642418i', '0.013055755639864134+0.010224251913449903i', '0.012780506779219061-0.01032096971828042i', '-0.01698539544684786-0.004548352676670059i', '-0.016962491656122458+0.004480805863341i', '0.01692487015763776-0.0014644674292922932i', '0.017113077569355553+0.0015728974096968704i', '-0.015302999195827751+0.007796634653028253i', '-0.015320209171170593-0.007819136194170087i', '0.01133311340114197-0.012932777704511678i', '0.011265391684377755+0.012857328465388215i', '-0.006254749426984171+0.01582140358586155i', '-0.006222158265875688-0.015981848055127694i', '-0.0010890564524260495-0.016777228498338875i', '-0.0010122718311515316+0.0170992836736931i', '0.005901661362111555+0.01647585222417855i', '0.006068943114251803-0.016712767686345338i', '-0.011629102668981445-0.011132910933798257i', '-0.011684659899693348+0.011113860501899234i', '0.01636227997621566+0.007732464827640534i', '0.016525530651467925-0.007756532218427016i', '-0.015817160764427813-0.0008623863274259819i', '-0.015969095423443602+0.0009968485325696774i', '0.01691752817102175-0.005653979265202752i', '0.01693153332834603+0.005723574108528126i', '-0.012276552944801053+0.010156697812479354i', '-0.012246317497925596-0.010203640733682504i', '0.008569475154164612-0.01610410344294657i', '0.008657921586077276+0.01615036886169327i', '-0.0027825412656850394+0.015077772176172982i', '-0.002813519732758507-0.015040726244644718i', '-0.005886951145398497-0.01718138075641685i', '-0.005933481956164638+0.01719457728332164i', '0.007910959048028913+0.01448437117577732i', '0.007940079735453034-0.014472084821097936i', '-0.014760719053701531-0.007508475601842091i', '-0.014749427137429744+0.007552642986228757i', '0.01689755785780657+0.004715229080795483i', '0.016809566063345723-0.004718035274094007i', '-0.015802710654601627+0.0035285953770526485i', '-0.015783527105942316-0.0035105195135580955i', '0.015102694777198718-0.008302417645422402i', '0.015125606463891786+0.008258512091757722i', '-0.009102898825326246+0.013381017809864351i', '-0.00913538504480442-0.013228898061620462i', '0.005598947699644214-0.016733372522338148i', '0.005580371621804248+0.016676176432362865i', '0.0018350343934440717+0.015518779405766528i', '0.0018358278233122916-0.015464585641131463i', '-0.008202881317286474-0.01577130937352199i', '-0.008048741193514383+0.015590583399400997i', '0.011781095224960663+0.010883320195464327i', '0.011527630825867802-0.010932564928688378i', '-0.01595027249813869-0.0056430192924029066i', '-0.01593116095464546+0.005563616264432501i', '0.016399444846542873-0.0001340454209290548i', '0.016571890772707862+0.0002239930861221047i', '-0.015305421837554723+0.006335639839616589i', '-0.01531771525375529-0.006358305266008887i', '0.011878616840627687-0.011569457840344702i', '0.011795977758592216+0.011513191518919563i', '-0.007227198906292837+0.01478626630120003i', '-0.0071977543010900424-0.014935260686678348i', '0.0002469424314421806-0.01618695441245411i', '0.00034349521925475246+0.01646389877325419i', '0.0045191537439027915+0.016334373629836335i', '0.004648490686719645-0.01656349021927873i', '-0.010269586201265283-0.011649844919420949i', '-0.01031894862269371+0.011639357951776487i', '0.01520713936219013+0.008594267156765645i', '0.015352278304378852-0.008620908192253578i', '-0.015209380594553866-0.0021151964246363397i', '-0.015340683389493542+0.0022499029529510772i', '0.016683885467651937-0.00422561261834116i', '0.01669299474875778+0.004279936637010092i', '-0.012679497165629466+0.00882397665274158i', '-0.012654710418124798-0.008862643424374716i', '0.009367382045684484-0.01487250062086385i', '0.00945629937769582+0.014914718514046094i', '-0.0039403724390886866+0.014432156359688642i', '-0.00396898405810521-0.014393870270542744i', '-0.004383309757569807-0.016858454206335922i', '-0.0044278995835937834+0.016872101293879678i', '0.006651068527456158+0.014703208739294524i', '0.006681293091310089-0.014689039209070744i', '-0.013529022288385525-0.008365651461723855i', '-0.01351742068567327+0.008409425466349246i', '0.016050700144423325+0.00576421980126722i', '0.01596146391118164-0.005763533629813664i', '-0.015499573112896521+0.0021249849175307423i', '-0.01548662789634514-0.002115375450734043i', '0.015231354595897494-0.00694122568117804i', '0.015249941776915847+0.006910851252852047i', '-0.009845863567811614+0.01216086944853964i', '-0.009858980114591728-0.012010626873110008i', '0.006612510749749334-0.01580751931067262i', '0.006578030982425001+0.015752422162347665i', '0.000498973019477697+0.015160347348840142i', '0.0005064309438526787-0.015114324518828257i', '-0.006768603685489323-0.015813359898609082i', '-0.006647962919493822+0.015627185155521944i', '0.010578999614756154+0.011496139518913859i', '0.010344481182867335-0.011505228412380927i', '-0.014985321470031919-0.006663352705080887i', '-0.014970685553626503+0.006574574177736513i', '0.015905959080763143+0.0011149287352030801i', '0.01606602937792906-0.0010414337504712837i', '-0.015307219215404524+0.0049671162570488064i', '-0.015315864445049399-0.0049899045372336255i', '0.012387733393663718-0.010293228754209176i', '0.01229332292388112+0.010253951857628393i', '-0.008143539225717714+0.013811070723375354i', '-0.008118062526302094-0.013950952356642406i', '0.001492347870893404-0.015630338815452734i', '0.0016063841719606425+0.015868743826087942i', '0.003211755107882505+0.016211934630870325i', '0.0033080224201401526-0.016434569820548518i', '-0.00898962198504901-0.012120048932539498i', '-0.009033693254924938+0.012116943301553872i', '0.01413292308159031+0.009414039372823424i', '0.014262789333878353-0.009443670942344669i', '-0.014623092505096878-0.003293528691860899i', '-0.01473602376117683+0.0034290712061047363i', '0.016474559169837054-0.00288513480030346i', '0.016479673768610606+0.0029271099285031223i', '-0.013049116994491683+0.007559440795934999i', '-0.013029003072723713-0.007591422285154137i', '0.010125355555009062-0.013730173520932025i', '0.010214751837536382+0.013767917809526489i', '-0.005040589995482819+0.01380126092817853i', '-0.005066783449166307-0.013761950003878075i', '-0.002989584374087556-0.016564136177685887i', '-0.0030322664261855092+0.01657833758996241i', '0.005437299300072994+0.014920245624582315i', '0.005468754699599995-0.014904937082615475i', '-0.012374791040607173-0.00915322594179934i', '-0.012362311761817846+0.009196612785040378i', '0.015256362451831089+0.006772859472205996i', '0.015166336086720918-0.006768144770214268i', '-0.01520318323662357+0.0008074405487800349i', '-0.015195479528756994-0.0008046409815857438i', '0.015363355010883522-0.005649410317111546i', '0.015378368959551192+0.00563010438902797i', '-0.010535290267380273+0.011006057854352568i', '-0.010530973641463395-0.010857843304541754i', '0.007583850797660864-0.014939291332285796i', '0.007535685926628163+0.014886805696554603i', '-0.0007643516260472244+0.014806361595031785i', '-0.0007512290748132033-0.014767457023401435i', '-0.0054178337951264754-0.01586296472864255i', '-0.005327094393415402+0.015672012896172337i', '0.00942981849078027+0.01207472978884637i', '0.00921224618930895-0.012048713572519573i', '-0.01407295588076067-0.007628270936567323i', '-0.014063204635178753+0.007531800974706358i', '0.015436122282920323+0.0023036600061517518i', '0.015586124053265073-0.002245205874081926i', '-0.015308596301677399+0.0036669986632094087i', '-0.015314448552714008-0.0036898761703332804i', '0.012870013247010322-0.00908161329750369i', '0.012765957772706796+0.009057773288580189i', '-0.009018967635561953+0.012879831106246076i', '-0.008997994129618103-0.013012189734299778i', '0.0026697684769553046-0.01509854388481946i', '0.0027997150800024573+0.015302934923828674i', '0.001958729341734862+0.01610477469581098i', '0.0020253316680610978-0.016321862712823636i', '-0.007767871716229814-0.012554010303366982i', '-0.007807250482999101+0.012557477571758676i', '0.013119722425298054+0.010204143802788432i', '0.013236279098367637-0.010237009937938652i', '-0.014050850965100614-0.0044173125109226025i', '-0.014146998392744373+0.004554107137348188i', '0.01628443650426737-0.0016096898125890907i', '0.01628617610719365+0.0016413196756443259i', '-0.013393024995564905+0.0063435854658928925i', '-0.013377069338205538-0.006369952030432598i', '0.010855043423700576-0.012655240659532753i', '0.01094488462595769+0.012688270715726117i', '-0.006100005911548079+0.013178082327582184i', '-0.006123806580314833-0.01313791286094851i', '-0.0016791222470156422-0.016292119628149443i', '-0.0017199570492283794+0.016306897483501278i', '0.00425353745150793+0.015137446247195666i', '0.004286332225675468-0.015121534743472706i', '-0.011277878066572197-0.009887270865991608i', '-0.011264103979043164+0.009930255476073058i', '0.014500864860955733+0.0077548530797395335i', '0.014410386708526908-0.0077457562484730845i', '-0.014910176927550776-0.00044676034259286124i', '-0.014907025340781782+0.00044386521067514735i', '0.01549937921531952-0.004406736840352782i', '0.01551131529977548+0.004396874934281947i', '-0.01118421843242412+0.009898237625581292i', '-0.011163783329316168-0.009752109240695583i', '0.008526877586010805-0.01411383697944036i', '0.008466501268343293+0.014064303829054228i', '-0.0019753014795946405+0.014453158776358388i', '-0.0019572092028259103-0.014420655245134008i', '-0.004127826611599654-0.015919637161921768i', '-0.004064575483718899+0.01572432473691657i', '0.008316838788022661+0.012628901300073256i', '0.008114833466065938-0.01257117882707938i', '-0.013198670737520228-0.008553350227206054i', '-0.013194036322671673+0.00845030329056753i', '0.014982954996923648+0.0034499661337078283i', '0.01512442387517611-0.003405608674076339i', '-0.015309698806101238+0.002415212639505601i', '-0.015313319695684197-0.0024381507397531216i', '0.0133333309082967-0.007915885224254647i', '0.013221022797787071+0.00790641380737453i', '-0.009866370479795002+0.011978982880440851i', '-0.009850248642922138-0.01210486816569582i', '0.0037978519222185564-0.01458403406932001i', '0.003942674715717395+0.01475742006658439i', '0.0007422924857547891+0.0160101386188189i', '0.0007815307360386492-0.016222333448177424i', '-0.006586281110987545-0.012959947501709992i', '-0.00662133816885159+0.012969458948044833i', '0.012151213688668581+0.010975370416930912i', '0.012255774236073338-0.011011600397596952i', '-0.013485795134683252-0.005503356672022099i', '-0.013566228803503486+0.00564170109373815i', '0.01610960890193571-0.0003801015389889445i', '0.01610838545592353+0.0004027135965204969i', '-0.013717280675336653+0.005159520373643492i', '-0.013705167291425291-0.005180956453482934i', '0.01156656244624895-0.011629876201664967i', '0.01165678260284636+0.011658071409236049i', '-0.007133240829973041+0.012555792270400483i', '-0.007154725592648556-0.012514894346477717i', '-0.0004302393458866585-0.01603744567137922i', '-0.00046930059400934754+0.016052759225877594i', '0.0030850747414704963+0.015356900811845105i', '0.003119305577563583-0.015340778487026776i', '-0.010221452517127923-0.01058062274003927i', '-0.010206068525281872+0.01062317307990473i', '0.013772748575453644+0.008722613863620866i', '0.013682076160537925-0.008708902129079203i', '-0.014617282621248978-0.0016566787215504114i', '-0.014618221623890593+0.0016488057008812833i', '0.01564042402033415-0.003195696267725208i', '0.01564958670904395+0.0031942775514716883i', '-0.011803284262393465+0.008821651276431697i', '-0.011767561729628812-0.008677602285454002i', '0.009454015063903295-0.013318720956838004i', '0.009382334344349166+0.013272365136555948i', '-0.003151311234754109+0.014097003167499983i', '-0.0031287142476331964-0.014070440042754206i', '-0.0028793323850603847-0.01598334990748964i', '-0.002842080752029505+0.015783895151263105i', '0.007225246753305797+0.013167006005593222i', '0.0070379139434292165-0.013079702111321874i', '-0.012350040618844996-0.0094519322729541i', '-0.012350635328479548+0.0093429914667372i', '0.014540347890565082+0.004569421930782219i', '0.014674235521765187-0.004538592217493798i', '-0.01531063658665728+0.0011943237867174522i', '-0.015312366621254958-0.0012172970523996254i', '0.013784459810501727-0.006779811548251233i', '0.013664730371708587+0.006784026790288923i', '-0.010697140272023078+0.0110965426573722i', '-0.010686091407846487-0.011216590231119497i', '0.004892565879553237-0.014080133929401106i', '0.005051664970773251+0.01422433070614237i', '-0.00045345453245421163+0.015926007145462183i', '-0.0004401571708866764-0.016133746976384086i', '-0.005428931860798583-0.013344478573682421i', '-0.005459855781285647+0.013359728322732394i', '0.011213515299716162+0.011737553056895667i', '0.011306881686913628-0.01177719474246723i', '-0.01292134263918128-0.006566437895187217i', '-0.012986710316433392+0.006706546922561677i', '0.015947037441602967+0.0008203675805481189i', '0.015943099887953616-0.0008059754323508743i', '-0.014026851774462917+0.0039919818720101224i', '-0.014018426574596426-0.004008866044959772i', '0.012269091253832978-0.010639017525972634i', '0.012359596314248135+0.010662339892466993i', '-0.008153573444567006+0.011927489296655928i', '-0.008172858600919389-0.011885965511953445i', '0.0007753433287726597-0.015796116692272544i', '0.0007379758873898995+0.015811870506077175i', '0.0019178875115028748+0.015580872090717249i', '0.0019536439366370645-0.01556482806580466i', '-0.00919087869234854-0.011243854330924334i', '-0.009173640558936769+0.011285922982406993i', '0.013062024542041362+0.0096878919727795i', '0.012971357489189199-0.00966941682325009i', '-0.014321193063414409-0.0028389725417951386i', '-0.014325940368641085+0.002826525513127339i', '0.01578778795239135-0.0020004510702683956i', '0.015794328941752193+0.0020069724651042113i', '-0.012401479745766164+0.007762174258695229i', '-0.012350899413901911-0.007620147134241157i', '0.010376881616855897-0.012543118032340636i', '0.010294341167848997+0.01250008120587182i', '-0.0043079647773539564+0.013733958264714081i', '-0.004281143769982176-0.013713084559801013i', '-0.0016553510818924505-0.016054470663696238i', '-0.0016433824762920601+0.01585092491265061i', '0.006141307962758011+0.013696489855887832i', '0.005968156590461585-0.013580676404687297i', '-0.011515969587789443-0.010335950314654611i', '-0.011521817153993502+0.010221447237128901i', '0.014102719463773027+0.005676243954678868i', '0.014229502000853322-0.005658695337526394i', '-0.015311499096732684-1.1492397682076605e-05i', '-0.01531149909673124-1.1492397462849191e-05i']
-;
-
-let freqArray = [];
-for (let i = 0; i < Math.round(fourierCoefficients.length / 2); i++) {
-  if (i == 0) {
-    freqArray.push(0);
-  } else {
-    freqArray.push(Math.sin(Math.PI / 2) * i);
-    freqArray.push(Math.sin((3 * Math.PI) / 2) * i);
-  }
-}
-
-console.log(freqArray);
-console.log(fourierCoefficients);
-
-/********** DRAWING SECTION **********/
-
-function setup() {
-  createCanvas(1880, 972);
-}
-
-function draw() {
-  background("#000000");
-  translate(940, 486);
-
-  const scaleCoeff = 4;
-
-  let intiX = 0;
-  let initY = 0;
-
-  let x = 0;
-  let y = 0;
-
-  for (let k = 0; k < fourierCoefficients.length; k++) {
-    let frequency = freqArray[k];
-
-    intiX = x;
-    initY = y;
-
-    x +=
-      scaleCoeff *
-      (math.re(math.complex(fourierCoefficients[k])) *
-        cos(2 * Math.PI * frequency * time) +
-        math.im(math.complex(fourierCoefficients[k])) *
-          sin(2 * Math.PI * frequency * time));
-    y +=
-      scaleCoeff *
-      (-1 *
-        math.im(math.complex(fourierCoefficients[k])) *
-        cos(2 * Math.PI * frequency * time) +
-        math.re(math.complex(fourierCoefficients[k])) *
-          sin(2 * Math.PI * frequency * time));
-
-    fill("#E9D2C0");
-    stroke("#E9D2C0");
-    line(x, y, intiX, initY);
-    ellipse(x, y, 5);
-  }
-
-  drawx.unshift(x);
-  drawy.unshift(y);
-
-  beginShape();
-  noFill();
-  for (let i = 0; i < drawx.length; i++) {
-    vertex(drawx[i], drawy[i]);
-  }
-  endShape();
-
-  point_limit = 30000;
-
-  if (drawx.length > point_limit) {
-    drawx.pop();
-  }
-  if (drawy.length > point_limit) {
-    drawy.pop();
-  }
-
-  time += 0.0007;
-}
